3-puts

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "main.h"
+#include <string.h>
 /**
 * _puts - prints a string.
 * @str: a string.
@@ -7,7 +8,7 @@
 */
 void _puts(char *str)
 {
-		int tmp = 0;
+		long unsigned int tmp = 0;
 
 		while (tmp < strlen(str))
 		{
